+++ dune/typetree/treepath.hh
@@ -12,7 +12,6 @@
 #include <type_traits>
 
 #include <dune/common/documentation.hh>
-#include <dune/common/version.hh>
 #include <dune/common/typetraits.hh>
 #include <dune/common/indices.hh>
 #include <dune/common/hybridutilities.hh>
@@ -69,27 +68,6 @@ namespace Dune {
         return std::integral_constant<std::size_t,v>();
       }
 
-      // these are helper functions that help triggering a deprecation warning
-      template<typename T>
-      [[deprecated("HybridTreePath index storage should be std::size_t or std::integral_constant<std::size_t,v>!\n"
-            "Using anything else is deprecated and will not possible after the 2.10 release.\n"
-            "It is adviced not to specify the template parameters expicitly,\n"
-            "but to use the helper functions `hybridTreePath` or `treePath`."
-            "These take care of converting indices to the appropriate storage.")]]
-      constexpr bool check_storage_type(MetaType<T>) {
-        return false;
-      }
-
-      // specialization of valid index type
-      template<std::size_t v>
-      constexpr bool check_storage_type(MetaType<std::integral_constant<std::size_t,v>>) {
-        return true;
-      }
-
-      // specialization of valid index type
-      constexpr bool check_storage_type(MetaType<std::size_t>) {
-        return true;
-      }
     }
 
     template<typename... T>
@@ -154,12 +132,9 @@ namespace Dune {
     class HybridTreePath
     {
 
-      // enable check for dune-typetree 2.10 and above
-#if DUNE_VERSION_GTE(TYPETREE,2,10)
       // make sure that all indices use std::size_t as the underlying number type
       static_assert((... && Impl::check_size_t<T>()),
         "HybridTreePath index storage must be std::size_t or std::integral_constant<std::size_t,v>");
-#endif
 
     public:
 
@@ -167,11 +142,7 @@ namespace Dune {
       using index_sequence = std::index_sequence_for<T...>;
 
       //! Default constructor
-      constexpr HybridTreePath()
-      {
-        [[maybe_unused]] constexpr bool check =
-          (... && Impl::check_storage_type(MetaType<T>()) );
-      }
+      constexpr HybridTreePath() = default;
 
       constexpr HybridTreePath(const HybridTreePath& tp) = default;
       constexpr HybridTreePath(HybridTreePath&& tp) = default;
@@ -182,20 +153,14 @@ namespace Dune {
       //! Constructor from a `std::tuple`
       explicit constexpr HybridTreePath(std::tuple<T...> t)
         : _data(t)
-      {
-        [[maybe_unused]] constexpr bool check =
-          (... && Impl::check_storage_type(MetaType<T>()) );
-      }
+      {}
 
       //! Constructor from arguments
       template<typename... U,
         typename std::enable_if_t<(sizeof...(T) > 0 && sizeof...(U) == sizeof...(T)),bool> = true>
       explicit constexpr HybridTreePath(U... t)
         : _data(t...) // we assume that all arguments are convertible to the types T...
-      {
-        [[maybe_unused]] constexpr bool check =
-          (... && Impl::check_storage_type(MetaType<T>()) );
-      }
+      {}
 
       //! Returns an index_sequence for enumerating the components of this HybridTreePath.
       [[nodiscard]] constexpr static index_sequence enumerate()